Applying change to logs.go, renaming internal grpc package to grpcutil
