+++ master/internal/grpcutil/auth.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcutil
 
 import (
 	"context"
