+++ master/internal/grpcutil/logging.go
@@ -1,4 +1,4 @@
-package grpc
+package grpcutil
 
 import (
 	"github.com/sirupsen/logrus"
